pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "big-data-factory"
version = "0.1.0"
description = "Production-oriented big data VM image factory."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1",
]

[project.scripts]
big-data-factory = "big_data_factory.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
